Prevent avoidable authentication failures that make the TPM refuse commands
The tcsd causes unnecessary authentication failures when a parent key (SRK)
with a password is used and a child key is supposed to be loaded into the
TPM. This patch avoids the failures by returning information about the
parent key to the caller rather than trying to load the key assuming the
parent key has no password (20 bytes of 0x0) and only returning information
about the parent key to the caller if the loading has failed. We use the
TCS_LOADKEY_INFO structure with all zeros as an indicator of a first call
and fill it out and return it to the client. On the client side we then
try to determine the password of the parent key and try another call with
either the parent key's password set or the parent key's password again
empty. This avoids one unnecessary authentication failure.

I noticed this issue when trying to develop an application that uses GnuTLS
tpmtool to create TPM keys and certtool for issuing certificates. I noticed
that with a hardware TPM I could only run my program once but could run it 5
times when using a vTPM. The vTPM seems to have a higher threshold for
engaging the lock out due to failed authentications than the hardware TPM does.
The program ran 5 times with the vTPM. My program itself does not cause
authentication failures but tcsd caused some, that this patch fixes, and
certtool itself causes some that would need to be 'somehow' fixed in certtool
itself. The improvement with this patch is so that the program now runs 5
times with a hardware TPM. After that the owner of the TPM has to issue the
TPM_ORD_ResetLockValue command.

The autentication failures are probably not so severe when the parent
key has no password. However, tpmtool (of GnuTLS) currently does not support
this and therefore we have to set an explicity password for the parent
key.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>